star bug solved
